+++ CHANGELOG.md
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Removed 
 
-- 
+- Removed warning messages from travelTimeMethodTimeOfConcentration
 
 ### Fixed